lms_api/requirements.txt
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.11.0
